Implement registration.
This allows a new user to register an account.  There are four required fields: first name, last name, email, password.  The user is required to enter the password a second time to confirm before the registration is accepted.  The link to the registration page is currently found on the login page - might want to add a link from the main page.