Work on action to buy stock
Implement stock lookup by ticker symbol.  Stock information is retrieved from Yahoo Finance.  Ticker symbols are currently assumed to be valid - will need to implement a check in the future. After a stock is retrieved, the information is shown along with a form to buy shares of the stock.  The user's bank balance is also shown.  Still need to implement the actual buying of the stock.