Implement ability to buy a stock.
A stock can be looked up using its ticker symbol.  The details of the stock are returned and the user is asked how many shares they want to buy.  The number of shares along with the symbol and purchase price are posted to buy_stock and added to the Stock table.  The user is redirected to the portfolio page.  Error checking is needed on the buy shares form to restrict the input to whole, positive numbers.