Updates to procedure to buy shares
Block purchase of shares if user does not have enough BrokrBucks.  Check to see if user already owns shares of the stock that they are trying to buy.  Inform user that they already own the stock and block the purchase (for now, this will be changed later). Require that user is logged in to view portfolio, buy_stock, and lookup_stock pages.